requirements.txt
pygame>=2.5.0
neat-python>=0.92
numpy>=1.24.0

psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
scikit-learn>=1.0.0